[리뷰 반영]
- blackCustomer, jdbcCustomer 삭제 후 SimpleCustomer로 둘 다 처리
- bufferedReader, bufferedWriter close() 코드 추가
