+++ src/main/java/com/prgrms/voucher_manager/customer/repository/JdbcCustomerRepository.java
@@ -1,7 +1,9 @@
 package com.prgrms.voucher_manager.customer.repository;
 
 import com.prgrms.voucher_manager.customer.Customer;
-import com.prgrms.voucher_manager.customer.JdbcCustomer;
+//import com.prgrms.voucher_manager.customer.JdbcCustomer;
+import com.prgrms.voucher_manager.customer.SimpleCustomer;
+import com.prgrms.voucher_manager.infra.Common;
 import com.prgrms.voucher_manager.wallet.Wallet;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -17,7 +19,7 @@ import java.time.LocalDateTime;
 import java.util.*;
 
 @Repository
-@Profile("dev")
+@Profile("default")
 public class JdbcCustomerRepository implements CustomerRepository {
 
     private static final Logger logger = LoggerFactory.getLogger(JdbcCustomerRepository.class);
@@ -29,14 +31,14 @@ public class JdbcCustomerRepository implements CustomerRepository {
     private static final String SELECT_COUNT_ALL_SQL = "SELECT COUNT(*) FROM customers";
 
     private static final RowMapper<Customer> customerRowMapper = (resultSet, i) -> {
-        UUID customerId = toUUID(resultSet.getBytes("customer_id"));
+        UUID customerId = Common.toUUID(resultSet.getBytes("customer_id"));
         String customerName = resultSet.getString("name");
         String email = resultSet.getString("email");
         LocalDateTime lastLoginAt =
                 resultSet.getTimestamp("last_login_at") != null ?
                 resultSet.getTimestamp("last_login_at").toLocalDateTime() : null;
         LocalDateTime createdAt = resultSet.getTimestamp("created_at").toLocalDateTime();
-        return new JdbcCustomer(customerId, customerName, email, lastLoginAt, createdAt);
+        return new SimpleCustomer(customerId, customerName, email, lastLoginAt, createdAt);
     };
 
 
@@ -101,10 +103,10 @@ public class JdbcCustomerRepository implements CustomerRepository {
         return hashMap;
     }
 
-    private static UUID toUUID(byte[] bytes) {
-        ByteBuffer byteBuffer = ByteBuffer.wrap(bytes);
-        return new UUID(byteBuffer.getLong(),byteBuffer.getLong());
-
-    }
+//    private static UUID toUUID(byte[] bytes) {
+//        ByteBuffer byteBuffer = ByteBuffer.wrap(bytes);
+//        return new UUID(byteBuffer.getLong(),byteBuffer.getLong());
+//
+//    }
 
 }
